backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
PyJWT
python-dotenv
sqlalchemy
pymysql
